standardize weights and epochs

+++ utils/utils.py
@@ -176,7 +176,7 @@ import torch
 def standardize_weights(dqn_weights, n_models):
     s_func = nn.Softmax(dim=0)
     means = [dqn_weights[0, cli*3] for cli in range(n_models)]
-    s_means = s_func(means)
+    s_means = s_func(torch.FloatTensor(means))
     s_std = [dqn_weights[0, cli*3+1]/10 for cli in range(n_models)]
     s_epochs = [math.ceil(dqn_weights[0,cli*3+1]*10) if math.ceil(dqn_weights[0,cli*3+1]*10) > 0 else 1 for cli in range(n_models)]
     assigned_priorities = [np.random.normal(s_means[i], s_std[i]) for i in range(n_models)]
